[chore] Make internal/featuregates into a module, move confmap fg there
