+++ internal/featuregates/featuregates.go
@@ -9,3 +9,11 @@ var UseUnifiedEnvVarExpansionRules = featuregate.GlobalRegistry().MustRegister("
 	featuregate.StageBeta,
 	featuregate.WithRegisterFromVersion("v0.103.0"),
 	featuregate.WithRegisterDescription("`${FOO}` will now be expanded as if it was `${env:FOO}` and no longer expands $ENV syntax. See https://github.com/open-telemetry/opentelemetry-collector/blob/main/docs/rfcs/env-vars.md for more details. When this feature gate is stable, expandconverter will be removed."))
+
+const StrictlyTypedInputID = "confmap.strictlyTypedInput"
+
+var StrictlyTypedInputGate = featuregate.GlobalRegistry().MustRegister(StrictlyTypedInputID,
+	featuregate.StageAlpha,
+	featuregate.WithRegisterFromVersion("v0.103.0"),
+	featuregate.WithRegisterDescription("Makes type casting rules during configuration unmarshaling stricter. See https://github.com/open-telemetry/opentelemetry-collector/blob/main/docs/rfcs/env-vars.md for more details."),
+)
