+++ processor/go.mod
@@ -65,3 +65,5 @@ replace go.opentelemetry.io/collector/pdata/testdata => ../pdata/testdata
 replace go.opentelemetry.io/collector/pdata/pprofile => ../pdata/pprofile
 
 replace go.opentelemetry.io/collector/config/configtelemetry => ../config/configtelemetry
+
+replace go.opentelemetry.io/collector/internal/featuregates => ../internal/featuregates
