+++ receiver/go.mod
@@ -65,3 +65,5 @@ retract v0.76.0 // Depends on retracted pdata v1.0.0-rc10 module
 replace go.opentelemetry.io/collector/config/configtelemetry => ../config/configtelemetry
 
 replace go.opentelemetry.io/collector/pdata/pprofile => ../pdata/pprofile
+
+replace go.opentelemetry.io/collector/internal/featuregates => ../internal/featuregates
