+++ versions.yaml
@@ -15,6 +15,7 @@ module-sets:
     version: v0.104.0
     modules:
       - go.opentelemetry.io/collector
+      - go.opentelemetry.io/collector/internal/featuregates
       - go.opentelemetry.io/collector/cmd/builder
       - go.opentelemetry.io/collector/cmd/mdatagen
       - go.opentelemetry.io/collector/component
